Integrate PosterGrid, fix closures injection on childs
ReccomendationView renamed to a more general PosterGrid
Reccomended movies are now actually visible  inside the details view of a movie.
Gotham  font has been added
Fixed issue with the injection of closures: whenever a child is instantiated by default without a closure that it has to propagate to its own childs the closure has to be provided with a didSet to propagate it  when later it is actually available.
Now a new detail page can be opened from the current detail page